Fix issue regarding warning about the two saved files
Mentioned in the UG that we are storing the appointment and clinic file separately. The change made manually to one file will not be automatically synchronised to the other file. Meaning If one of the files crashes, the other would not be affected. So users need to be more cautious when editing those two files directly.